[not verified] Widget visibility: Move filter referer check
Previously registration of the front-end hooks was not done in the
referer was the widgets.php admin page. This prevented the API from
running the filters which made filtered out widgets uneditable.

This also had the effect of running the filters if you clicked from the
widgets page to e.g. the homepage, which would look confusing to users.

This has been solved by changing the check so that filters aren't ran if
on the REST API AND the referer is the widgets.php admin page.