tools: version-packages.sh should update indirect deps too (#20363)
It was updating the direct dependencies, but leaving the indirect
dependencies at monorepo versions. This resulted in us releasing Jetpack
9.8 with a dev-master version of automattic/jetpack-password-checker.

Since we now have that extra `composer update` call at the end to update
all the packages, we can pass --no-update to the `composer require`
calls to save all the updates until the end.