tools: Add the ability to do `@since $$next-version$$` (#20386)
And similarly for `@deprecated`. More specifically, if you write any of

- `@since $$next-version$$`
- `@deprecated $$next-version$$`
- `@deprecated since $$next-version$$`

then the `$$next-version$$` will be replaced with the appropriate
version number on release (using `tools/changelogger-release.sh`).

The appropriate alpha version will also be used when for the mirror repo
(via `.github/files/build-all-projects.sh`).

Using the string `$$next-version$$` will result in an error during the
build.