Provide `isRTL` as an export in the `i18n-to-php.js` webpack extension
This extension is used in place of all `@wordpress/i18n` imports throughout the js codebase. Relatively recently, the `i18n` package stated to export an `isRTL` function, and a lot of consumers use it. Since in Jetpack this package is replaced by this extension - that doesn't export `isRTL` - they were then breaking the build.