mirrors: Use a token secret in the autotagger
We're going to want to trigger some other actions on tag, but taggings
with the default GITHUB_TOKEN specifically do not trigger other actions.

So we'll need to add a token secret to all the mirrors using autotagger.
Added a tool to do mass updates of repo secrets to help with that.