tools: Fix package.json versioning in changelogger-release.sh
When changelogger-release.sh releases a JS package, other packages that
depend on it may in turn need updating. Have it run
check-intra-monorepo-deps.sh to do so.

But when it runs check-intra-monorepo-deps.sh, that script shouldn't
create new change entries for the packages that are being released. So
give it options to allow it to skip the change entry creation.

Also replace-next-version-tag.sh was giving weird `grep` errors because
of deleted change entry files. Avoid that.