REST API: Adds 'i_replied' field to comments in comments endpoint responses.
Summary:
This adds a new field to comments returned by comments endpoints.  The new field indicates whether the current user has replied to a specific comment.
The field is a simple bool rather than an int for the ID of the reply since there could be more than one.

More context in this thread: https://[private link]#comment-3471

Test Plan:
Sandbox this change and visit the following endpoints for sites where you have replied to comments.  Confirm that `i_replied` is `true` for comments to which you have replied and is `false` otherwise.

/sites/%s/comments
/sites/%s/comments/%d
/sites/%s/posts/%d/replies

Reviewers: emilylaguna

Reviewed By: emilylaguna

Tags: #touches_jetpack_files

Differential Revision: D64353-code

This commit syncs r229758-wpcom.