Admin: check for offline mode when enabling Photon via main toggle
The Site Accelerator toggle is a custom toggle that works a bit differently from the standard ModuleToggle component. Since it handles 2 modules at once (Asset CDN and Photon), it needs to check if the 2 modules can actually be toggled on before to activate them. Until now, we checked if the Photon module had an override set via a filter, but we did not check if it was rendered unavailable via Offline mode. This changes that.

This was brought up in https://github.com/Automattic/jetpack/pull/20534#issuecomment-892836173