Related Posts: process content in post titles
Fixes #20551

Related Posts' titles need to be quite simple, to avoid display issues in the Related Post grid and to avoid displaying HTML tags in the title attribute used in the related posts anchor. That said, we should be able to allow shortcodes that add simple text to post titles to work.

Let's consequently continue to strip all tags, but let's process the post titles through the the_title filter so third-party plugins can use it.

This may also address #2535.