required-review: Fix handling of multiple paths and negation
Passing multiple path-patterns with some of the non-first patterns being
negated gave unexpected results. For example,

    paths:
      - '!a'
      - '!b'

would match everything including `a` and `b`, since the `!a` matches `b`
and the `!b` matches `a`.

What we want is for a non-first negated pattern to only exclude paths
that were matched by an earlier pattern, it should not add new paths.
That's the behavior already implied by README.md.