[not verified] Add filter to override critical css proxy skipped hosts
While generating critical CSS, the CSS files need to be proxied through a the plugin to avoid CORS restrictions. Previously the proxy was skipped if the server hostname matched the request. Or if there was no hostname provided in the url. Now, it has a filter hook to allow more control.