Add a webpack.config.js to lazy-images
Webpack 5 breaks if you pass `argv.entry` through as `entry` from your
config function. Webpack 4 was ok with that.

I have no idea whether that's a bug in webpack 5 for breaking that, or
calypso-build for passing it through instead of letting webpack handle
it. Either way, this works around it.

Also of note is that calypso-build now declares webpack as a peer dep,
so lazy-images needs to depend on it to provide the peer.