Update webpack globally  (#20789)
* Globally update to latest webpack, weback-cli, calypso-build

* Changelogs

* Get lazy-images build working, and bring storybook webpack up to date

* Changelog

* Revert removal of calypso-build in lazy-images

It added a @todo saying to make it use calypso-build again, which I
figured out how to do.

* Add a webpack.config.js to lazy-images

Webpack 5 breaks if you pass `argv.entry` through as `entry` from your
config function. Webpack 4 was ok with that.

I have no idea whether that's a bug in webpack 5 for breaking that, or
calypso-build for passing it through instead of letting webpack handle
it. Either way, this works around it.

Also of note is that calypso-build now declares webpack as a peer dep,
so lazy-images needs to depend on it to provide the peer.

* lazy-images: Update ecmaVersion

No more IE11 => newer ecmaVersion

* Use @babel/preset-env to emit valid es2015

* Revert move to es9 validation

* Simplify

This works for the moment thanks to `@wordpress/browserslist-config`'s
"> 1%" including Opera Mini.

We'll have to update the "validate-es5" check if we update the JS
sources to modern standards, or that usage drops below 1%.

Co-authored-by: Brad Jorsch <brad.jorsch@automattic.com>