Use Node 16.7.0 in tooling
To match the version of Node used in Calypso.

This shouldn't change the behavior of any of the code itself.
The main "risk" as far as incompatibility is if future PRs to
eslint-changed start using Node 16 features. Everything else that's
published is targeted at browser environments or is an action where our
Node 14 was already past GitHub's Node 12.