storybook: Use webpack 5 everywhere (#20891)
They refactored everything to use either 4 or 5, but left a lot of
webpack 4 deps in various subpackages. Either unused, or using
functionality that wasn't changed between 4 and 5.

This PR adds overrides to force webpack 5 in place of 4. That removes a
bunch of webpack 4 deps on chokidar 2, which is itself deprecated and
depends on a version of glob-parent that `pnpm audit` complains about.