js libs: Set `exports` in `package.json` for renovate
Renovate decides whether a package is a library or not by looking for
the presence of a `main` or `exports` key in `package.json` (and that
it's not `private`). If it's not considered a library, renovate will try
to pin dependencies in the package.

So, for our JS libraries, set `exports` so renovate will consider them
as libraries. Also add a CI lint that will try to catch this sort of
thing in the future.