renovate: Fix config
Changes in #20917 resulted in some errors about "Each packageRule must
contain at least one non-match* or non-exclude* field". Apparently just
an `extends` isn't enough.

Also, even though it doesn't complain, the `extends` value is supposed
to be an array. So make it so. And `supportPolicy` has been deleted, so
remove it.