Contact Form: adjust wording for filter description.
One can use this filter to automatically set up cc or bcc for all contact forms on the site, based on data submitted by site visitors.
See the description in #20912 for an example.

Let's make it clear that by doing so, you're allowing visitors to send emails to any email address they want.