tools: Fix check-intra-monorepo-deps.sh in CI mode (#20955)
`pnpm install` defaults to `--frozen-lockfile` when running in GH Actions,
which is specifically what we don't want when the renovate helper runs
`check-intra-monorepo-deps.sh -u`. So have it detect when it's running
in CI and pass `--no-frozen-lockfile` in that case (and also omit the
`--silent` that made this so hard to debug).