Fix git hooks when Jetpack is a git submodule
The code was relying on the `.git` directory being in the repository root, but this is not always the case. For example, when Jetpack is a submodule within another git repository, `.git` is a file that points to the parent repository's `.git/modules/jetpack` folder.
More information about how `.git` works in the context of a git submodule can be found here: https://github.com/git/git/blob/master/Documentation/RelNotes/1.7.8.txt#L109-L114

The solution here is to use `git rev-parse --git-dir` to find the correct location of the `.git` folder.