jetpack: Use `[contenthash]` in Webpack chunk names
This will improve the long-term caching of the files by only busting
caches of the files that have changed rather than busting the cache of
every associated file when one does.

This should also help reduce the instances of the wpcom TeamCity build
deciding to change all the hashes, although completing that will also
require a fix for WordPress/gutenberg#34660.