Fix check dev env script when Jetpack is submodule
Similar to https://github.com/Automattic/jetpack/pull/20971, this PR fixes an issue that arises when Jetpack is a submodule of another repository.
Here, when Jetpack is a submodule, the `tools/check-development-environment.sh` script reports not being inside a git checkout and incorrectly prompts the user to clone the repository.

Switching from inspecting `.git` to inspecting the output of `git rev-parse --git-dir` was the solution used in the previous PR and is what this change uses as well.