repo-gardening: Use `pull_request_target` for all tasks (#21078)
This will let them run for PRs from forked repos. Also it'll reduce the
number of workflow runs, since we won't have runs for both
`pull_request` and `pull_request_target`.

The main drawback is some extra complication for the checkDescription
task, as that looks at the filesystem to determine whether changelogger
was properly used.