Docker: Fix shebang printing from `wp` (#21116)
The way `psysh` is being installed causes `wp` to print `#!/usr/bin/env
php` every time it is run.

Also install.sh prints a weird error (and probably fails to detect
whether WP is installed) because of this plus another bug.

Instead, let's install psysh as a library and require that in `wp`
instead. And may as well do so once globally, instead of blindly
appending the line every startup to /var/www/html/wp-cli.yml.