docker: Use the same PHP and Composer versions as CI
We can have the Jetpack CLI read the `.github/versions.sh` file to
determine the needed versions, and install those instead of hard-coding
to PHP 7.4 and whichever old version of composer is in the version of
Ubuntu being used.

I also took the opportunity to rearrange a few other things in the
image, primarily moving the config files from /tmp to
/var/lib/jetpack-config where they're less likely to be accidentally
messed with.