Test scripts need to decide which version of PHPUnit to run
So WP backported the "require polyfills" logic to 5.7 and 5.8 (but not
5.8.1), but they didn't backport the "stop requiring an outdated version
of PHPUnit" logic. 🙄

So our test scripts need to check whether the current version of
WordPress is in that situation, and if so run PHPUnit 7 instead of
whatever later version got pulled in by yoast/phpunit-polyfills.

Also we need to use a PHPUnit 7 phar when doing that, as the
globally-composer-installed version we had been using winds up getting
confused and still loads a bunch of classes from Jetpack's vendor
directory while phar versions load all the deps from the phar.