Remove references to global phpunit
It's not going to work right, so let's stop providing it in CI and
Docker, and stop telling people to provide it in the docs.

Note this leaves the `jetpack docker phpunit` command in an explicitly
broken state. I'm going to leave fixing that for a followup. The thing
to do is probably to have it call the new `composer phpunit` script
added in this commit.