Fix composer `.autoload.files`
These files are loaded as soon as the autoloader is, which might be
early in PHPUnit's execution before anything gets around to loading
WordPress.

The files in packages/backup and packages/compat were just trying to
register actions and filters. If they were loaded by the autoloader
before WordPress's plugin API was initialized, we can drop data into
`$wp_filter` for `WP_Hook::build_preinitialized_hooks()` to find later.

The one in packages/connection was trying to load some WordPress classes
that might otherwise not be loaded. It seems more straightforward to do
that in the three places it might be needed instead of doing it at
autoload time.