Move the identity-crisis files from the identity-crisis pacckage to the connection package
There is circular dependency between the identity-crisis and connection packages. Since all connections
should be able to detect an identity crisis, and an identity crisis depends on having a connection, move
the identity crisis files to the connection package.

Also remove the identity-crisis dependencies from the sync package and the Jetpack plugin.