CI: Run tests with PHP 8.1, experimentally
PHP 8.1 is getting close to release. Let's run tests with it so we can
start cleaning stuff up.

This is certainly going to fail a lot of stuff for the moment, as
WordPress itself is [still working on this][1]. But we may as well get a
head start on it.

[1]: https://core.trac.wordpress.org/ticket/53635