ci: Fix composer monorepo repo entry in copied plugin (#21218)
The maybe-downgrade-phpunit logic added in #21157 was just deleting the
monorepo repo entry, which was causing it to re-download all the
monorepo deps from packagist. That's wrong, it should continue to use
the monorepo packages that were checked out.

To avoid that, when the setup-wordpress-env script copies the plugin
into place it needs to update that repo entry in composer.json to point
to the correct path (and set options.symlink = false to match the logic
used originally, and versions because otherwise composer gets confused).