Avoid calypso-build's babel-plugin-optimize-i18n plugin
It's not clear what exactly it's intended to do, but in the context of
Jetpack is seems to brokenly result in calls to WordPress's i18n
functions (`__` and so on) to be minified away.

Excluding it brings us back to the situation before the upgrade to
calypso-build 9.0.0, which should fix #21204.