Jetpack: Exclude `vendor/` from coverage (#21238)
Something in #21218 seems to have changed the structure of the `vendor/` in
a way that makes PHPUnit take a very long time to generate the code coverage
report. Since there's no point in it including coverage for `vendor/` (as
nothing in there is part of Jetpack itself; even our monorepo modules should
have their own tests covering them), let's exclude it.