Hide image block's caption and settings
As the image block's caption component and settings are not a part of the Tiled Gallery block's design, they're hidden with this commit. This is done by adding an "hideSettings" context and setting it to true. This is passed back to the image block, which is set to hide the unwanted settings when this context is detected.