Post List: Add the Share post action to Posts and Pages (#21300)
* [not verified] Post List: Add the Share post action to Posts and Pages

This adds the share action to the post list for the 'post' and 'page'
post types. It uses the query parameter to ensure the Jetpack sidebar is
open when the editor loads.

Determining exactly when it's appropriate to have the Share action
available is tricky and based on context. For now we are using a filter,
so the exact logic can be applied as the Post List package is configured
for each environment.

* [not verified] Fix the unit tests

* Break up the long if statement

* Update package versions