[not verified] Post List: Return null when the attachment URL is unavailable
In testing the package on wpcom I had some pages where there was an
attachment ID in the post content, but that ID wasn't available on the
site.

This meant that the thumbnail was be being returned with empty strings
for the `src` and `alt` attributes, and a broken image was displayed.

This change updates the checks to ensure we have a valid thumbnail URL,
and if not, returns `null`