Jetpack: activate default modules when the site is already connected
Currently, the default Jetpack modules are activated when the site is registered and when the
first user is authorized. However, there are other plugins that can be used to register the site
and authorize the user. If the site is already connected via another plugin when the Jetpack
plugin is activated, the default modules will not be activated.

Fix this by adding a new method, `Jetpack::activate_default_modules_when_already_connected`, which will
do the following:

 * Activete the modules in the `jetpack_active_modules` option if the option exists. If the site previously
   had Jetpack active, this will keep the user's previous modules choices activated.
 * If the option does not exist, activate the site-only connection default modules. If a user is
   connected, also activate the user-connection default modules.