Add tooling for weekly Atomic releases
After much discussion (p1634048495196700-slack-C01U2KGS2PQ), we've
decided that the weekly Atomic releases for Jetpack should be versioned
like "10.3-a.1", "10.3-a.3", and so on (odd numbers), while the master
alpha versions should be like "10.3-a.0", "10.3-a.2", and so on.

This updates the tooling to make that happen. In several places we need
to decide whether to pass `--prerelease=alpha` or `--prerelease=a.0` to
changelogger, and `changelogger-release.sh` needs to have a flag to
specify that one of these Atomic releases is being created.