changelogger: Fix `version next` when current version is a prerelease
Currently, if the current version in the changelog is "1.2.3-alpha",
a patch bump would want the next version to be "1.2.4" where a user
would probably expect to be releasing "1.2.3". It's even worse if you
pass say `--prerelease=alpha.2`, it would want to go to to
"1.2.4-alpha.2" instead of "1.2.3-alpha.2".

So, let's DWIM here: If the latest version is a prerelease, find the
most recent non-prelease version and bump that instead. But if it turns
out that produces something less then the actual current version, try
just updating the prerelease of the current version. And if that too is
less (e.g. "1.2.3-beta" with --prerelease=alpha), then do a patch bump
from the current version.

Unfortunately to do this we need to update the VersioningPlugin
interface to provide a way to do "just updating the prerelease of the
current version", which makes for a major version bump. Sigh.