changelogger: Fix version next for a -beta to the following -alpha
In the previous PR, we had it look at the latest release version as well
as the latest version when determining the next version. That logic
seems to have a small bug. Consider the following:

* Latest release was 1.2.0.
* Latest version is 1.3.0-beta, in preparation for a 1.3.0 release.
* We're on the dev branch, where there's another minor-significance
  change pending, and wanting the next -alpha version.

The current logic produces 1.3.1-alpha, while the actual next release
(assuming that minor bump isn't cherry-picked back into the beta) will
be 1.4.0-alpha.