docker: Add script to create "fake" translations for testing i18n
This adds a script, intended to be run inside the Jetpack Docker
environment, to create plugin translations for two fake languages.

- en_piglatin: All messages are translated into Pig Latin.
- en_rtl: All messages will display backwards, and the interface will be
  right-to-left.

Once you have run the script, in WordPress Admin > Settings > General
you should see the above two languages as options for "Site Language".

Note WordPress itself is not translated as described, the translations
are only created for messages in the specified plugin.