docker: Add jq to the Docker env (#21424)
We use it for a bunch of our tooling, to the point where
tools/check-development-environment.sh requires it. It should be present
in the docker env too.

This increases the size of the image from 665.3M to 665.7M.

Also, every time I look at the Dockerfile, I wonder whether
"ubuntu:focal" is the up to date version or not. GitHub's Actions
environment refers to it as "20.04", which seems clearer to me.

This has no effect on the image at all, as "20.04" and "focal" are tags
for the same image.