cli: Fix generation of js-packages to not include a "require-dev " (#21436)
Not sure why this was even there in the first place. We already get the
dep on changelogger from the skeleton file. If it was supposed to be
removing the yoast/phpunit-polyfills dep, that would be better done by
an overriding skeleton file or a use of the `delete` operator so as to
avoid having another place where the changelogger version is specified.