Identity Crisis: Remove the unused jetpack_idc_option transient
The `jetpack_idc_option` transient is set in Identity_Crisis::get_sync_error_idc_option, but it's
never used. Let's remove it.

We also display the value of the transient in the IDC Simulator module in the Debug Helper plugin. Since
the transient will no longer be set, remove it from the IDC simulator UI.