Debug Helper IDC Simulator: Add the ability to spoof the home url
The Debug Helper plugin's IDC Simulator module allows the user to spoof the siteurl value. Let's
add the ability to also spoof the home value. To do this, add two checkboxes that allow the user
to choose whether to spoof the siteurl and/or home values. These checkboxes should be checked by
default because the most common use case will be to spoof both values.