Debug Helper IDC Simulator: add setting to enable/disable Sync
When testing the IDC flow with request to REST endpoints, it can be helpful to disable
Sync requests. Sync requests use a separate IDC flow and could make changes during testing.

To make testing easier, add a setting which allows the user to enable and disable Sync. The
setting works by setting the `option_jetpack_sync_settings_disable` option to true when the
user selects to disable Sync.