Replaced += (addition operator) to array_merge function call.
This will fix the issue when any two/more arrays index value is same.
Example of the same issue can be found in the issue comment.
NOTE: If the input arrays have the same string keys,
then the later value for that key will overwrite the previous one.
If, however, the arrays contain numeric keys, the later value will
not overwrite the original value, but will be appended.