Debug Helper IDC Simulator: display the last five remote requests (#21534)
When testing IDCs, it can be useful to look at the url to verify that
the IDC query args are added when expectd. It's also useful to see if
WPCOM responded with the expected idc_detected and migrated_for_idc
values. Display the last 5 request urls and the IDC response on the
IDC Simulator page.