webpack-config: Fork calypso-build's mini-css-with-rtl plugin
And clean it up with some better formatting, better input validation,
and a better way to override the default "isRTL" expression.

Also, split the "RtlCssPlugins" into two now that StandardPlugins looks
to be the most common way of getting plugins. That makes it more
straightforward to configure the two plugins.