eslint-changed: Use API for running ESLint
We had been using ESLint's API just to call `CLIEngine.getFormatter()`,
which they deprecated in 7.0.0 and removed in 8.0.0. There's no direct
replacement; they have `loadFormatter()`, but not as a static method.

Since we're having to instantiate ESLint now anyway to get the
formatter, let's switch over to its API to do the actual linting too
instead of shelling out.