Post List: Fix the stretched thumbnail images.
When we take the thumbnail image from the post content, it might not be
the correct aspect ratio so, as it is displayed in a square, it appears
stretched and distorted.

This fixes that by applying `object-fit` to the image. This CSS isn't
supported in IE, but I believe we're no longer supporting IE, and it's a
visual glitch, that doesn't otherwise break the experience.