Sync: do not run an initial sync if an initial sync has already been run
Do not run an initial sync when these two conditions are true:
 - The site is not registering.
 - An initial sync has been run before.

This change will allow plugins to run an initial sync outside of the site
registration process. For example, if a plugin starts using sync after an update,
it can run an initial sync to get the cache site up to date. However, this isn't
necessary if another plugin has already run an initial sync.