Use it in Jetpack
Jetpack has a bunch of different ways scripts get built. This only
touches the ones that are built with our webpack config.

Further, this doesn't touch admin.js or the blocks' view.js. For the
former, the JS and CSS are being registered in different places and the
CSS is being enqueued twice for some reason. For the latter, there's a
bunch of extra logic for handling previews that I didn't want to mess
with in this PR.