Status: add a function_exists check before calling wp_get_environment_type (#21682)
The wp_get_environment_type function was added in WordPress 5.5. Consumer plugins
may not be enforcing a minimum WP version of 5.5. Using this package on sites with
WP <5.5 will cause a fatal error if wp_get_environnment_type is called..

We intend to define minimum WP requirements for the packages in the future.

In the meantime, check that the wp_get_environment_type function exists before calling it.