[not verified] Post List: Fix the response after saving a Quick Edit.
The customizations that to the post list screen were being made on hooks
that only ran as the page/screen loaded. If a quick edit was made, the
table row returned in the result didn't include the necessary
customizations and would break the table.

There appears to be a WordPress bug where the post actions are not
hidden after a quick edit too. I also noticed the Copy post action was
missing, which might be another issue in Jetpack.